LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_fractional_coordinates_untrusted-expected.txt


FAIL PointerEvent.pointerdown.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerdown.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerdown.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerdown.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerdown.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerdown.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerdown.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerdown.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerup.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerup.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerup.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerup.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerup.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerup.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerup.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerup.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointermove.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointermove.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointermove.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointermove.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointermove.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointermove.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointermove.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointermove.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointercancel.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointercancel.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointercancel.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointercancel.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointercancel.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointercancel.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointercancel.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointercancel.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.gotpointercapture.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.gotpointercapture.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.gotpointercapture.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.gotpointercapture.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.gotpointercapture.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.gotpointercapture.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.gotpointercapture.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.gotpointercapture.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.lostpointercapture.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.lostpointercapture.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.lostpointercapture.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.lostpointercapture.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.lostpointercapture.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.lostpointercapture.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.lostpointercapture.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.lostpointercapture.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerrawupdate.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerrawupdate.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerrawupdate.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerrawupdate.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerrawupdate.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerrawupdate.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerrawupdate.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerrawupdate.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerover.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerover.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerover.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerover.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerover.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerover.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerover.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerover.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerout.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerout.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerout.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerout.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerout.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerout.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerout.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerout.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerenter.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerenter.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerenter.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerenter.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerenter.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerenter.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerenter.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerenter.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerleave.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.pointerleave.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.pointerleave.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerleave.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerleave.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerleave.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.pointerleave.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.pointerleave.offsetY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.foo.screenX assert_equals: expected 0.111111111111 but got 0
FAIL PointerEvent.foo.screenY assert_equals: expected 0.222222222222 but got 0
FAIL PointerEvent.foo.clientX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.foo.clientY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.foo.pageX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.foo.pageY assert_equals: expected 0.444444444444 but got 0
FAIL PointerEvent.foo.offsetX assert_equals: expected 0.333333333333 but got 0
FAIL PointerEvent.foo.offsetY assert_equals: expected 0.444444444444 but got 0
PASS 0.0 - 0.5: PointerEvent.click.screenX
PASS 0.0 - 0.5: PointerEvent.click.screenY
PASS 0.0 - 0.5: PointerEvent.click.clientX
PASS 0.0 - 0.5: PointerEvent.click.clientY
PASS 0.0 - 0.5: PointerEvent.click.pageX
PASS 0.0 - 0.5: PointerEvent.click.pageY
PASS 0.0 - 0.5: PointerEvent.click.offsetX
PASS 0.0 - 0.5: PointerEvent.click.offsetY
PASS 0.5 - 1.0: PointerEvent.click.screenX
PASS 0.5 - 1.0: PointerEvent.click.screenY
PASS 0.5 - 1.0: PointerEvent.click.clientX
PASS 0.5 - 1.0: PointerEvent.click.clientY
PASS 0.5 - 1.0: PointerEvent.click.pageX
PASS 0.5 - 1.0: PointerEvent.click.pageY
PASS 0.5 - 1.0: PointerEvent.click.offsetX
PASS 0.5 - 1.0: PointerEvent.click.offsetY
PASS 1.0 - 1.5: PointerEvent.click.screenX
PASS 1.0 - 1.5: PointerEvent.click.screenY
PASS 1.0 - 1.5: PointerEvent.click.clientX
PASS 1.0 - 1.5: PointerEvent.click.clientY
PASS 1.0 - 1.5: PointerEvent.click.pageX
PASS 1.0 - 1.5: PointerEvent.click.pageY
PASS 1.0 - 1.5: PointerEvent.click.offsetX
PASS 1.0 - 1.5: PointerEvent.click.offsetY
FAIL -0.5 - 0.0: PointerEvent.click.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.click.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.click.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: PointerEvent.click.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.click.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: PointerEvent.auxclick.screenX
PASS 0.0 - 0.5: PointerEvent.auxclick.screenY
PASS 0.0 - 0.5: PointerEvent.auxclick.clientX
PASS 0.0 - 0.5: PointerEvent.auxclick.clientY
PASS 0.0 - 0.5: PointerEvent.auxclick.pageX
PASS 0.0 - 0.5: PointerEvent.auxclick.pageY
PASS 0.0 - 0.5: PointerEvent.auxclick.offsetX
PASS 0.0 - 0.5: PointerEvent.auxclick.offsetY
PASS 0.5 - 1.0: PointerEvent.auxclick.screenX
PASS 0.5 - 1.0: PointerEvent.auxclick.screenY
PASS 0.5 - 1.0: PointerEvent.auxclick.clientX
PASS 0.5 - 1.0: PointerEvent.auxclick.clientY
PASS 0.5 - 1.0: PointerEvent.auxclick.pageX
PASS 0.5 - 1.0: PointerEvent.auxclick.pageY
PASS 0.5 - 1.0: PointerEvent.auxclick.offsetX
PASS 0.5 - 1.0: PointerEvent.auxclick.offsetY
PASS 1.0 - 1.5: PointerEvent.auxclick.screenX
PASS 1.0 - 1.5: PointerEvent.auxclick.screenY
PASS 1.0 - 1.5: PointerEvent.auxclick.clientX
PASS 1.0 - 1.5: PointerEvent.auxclick.clientY
PASS 1.0 - 1.5: PointerEvent.auxclick.pageX
PASS 1.0 - 1.5: PointerEvent.auxclick.pageY
PASS 1.0 - 1.5: PointerEvent.auxclick.offsetX
PASS 1.0 - 1.5: PointerEvent.auxclick.offsetY
FAIL -0.5 - 0.0: PointerEvent.auxclick.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.auxclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.auxclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: PointerEvent.auxclick.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.auxclick.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: PointerEvent.contextmenu.screenX
PASS 0.0 - 0.5: PointerEvent.contextmenu.screenY
PASS 0.0 - 0.5: PointerEvent.contextmenu.clientX
PASS 0.0 - 0.5: PointerEvent.contextmenu.clientY
PASS 0.0 - 0.5: PointerEvent.contextmenu.pageX
PASS 0.0 - 0.5: PointerEvent.contextmenu.pageY
PASS 0.0 - 0.5: PointerEvent.contextmenu.offsetX
PASS 0.0 - 0.5: PointerEvent.contextmenu.offsetY
PASS 0.5 - 1.0: PointerEvent.contextmenu.screenX
PASS 0.5 - 1.0: PointerEvent.contextmenu.screenY
PASS 0.5 - 1.0: PointerEvent.contextmenu.clientX
PASS 0.5 - 1.0: PointerEvent.contextmenu.clientY
PASS 0.5 - 1.0: PointerEvent.contextmenu.pageX
PASS 0.5 - 1.0: PointerEvent.contextmenu.pageY
PASS 0.5 - 1.0: PointerEvent.contextmenu.offsetX
PASS 0.5 - 1.0: PointerEvent.contextmenu.offsetY
PASS 1.0 - 1.5: PointerEvent.contextmenu.screenX
PASS 1.0 - 1.5: PointerEvent.contextmenu.screenY
PASS 1.0 - 1.5: PointerEvent.contextmenu.clientX
PASS 1.0 - 1.5: PointerEvent.contextmenu.clientY
PASS 1.0 - 1.5: PointerEvent.contextmenu.pageX
PASS 1.0 - 1.5: PointerEvent.contextmenu.pageY
PASS 1.0 - 1.5: PointerEvent.contextmenu.offsetX
PASS 1.0 - 1.5: PointerEvent.contextmenu.offsetY
FAIL -0.5 - 0.0: PointerEvent.contextmenu.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: PointerEvent.contextmenu.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: PointerEvent.contextmenu.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: PointerEvent.contextmenu.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: PointerEvent.contextmenu.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.click.screenX
PASS 0.0 - 0.5: MouseEvent.click.screenY
PASS 0.0 - 0.5: MouseEvent.click.clientX
PASS 0.0 - 0.5: MouseEvent.click.clientY
PASS 0.0 - 0.5: MouseEvent.click.pageX
PASS 0.0 - 0.5: MouseEvent.click.pageY
PASS 0.0 - 0.5: MouseEvent.click.offsetX
PASS 0.0 - 0.5: MouseEvent.click.offsetY
PASS 0.5 - 1.0: MouseEvent.click.screenX
PASS 0.5 - 1.0: MouseEvent.click.screenY
PASS 0.5 - 1.0: MouseEvent.click.clientX
PASS 0.5 - 1.0: MouseEvent.click.clientY
PASS 0.5 - 1.0: MouseEvent.click.pageX
PASS 0.5 - 1.0: MouseEvent.click.pageY
PASS 0.5 - 1.0: MouseEvent.click.offsetX
PASS 0.5 - 1.0: MouseEvent.click.offsetY
PASS 1.0 - 1.5: MouseEvent.click.screenX
PASS 1.0 - 1.5: MouseEvent.click.screenY
PASS 1.0 - 1.5: MouseEvent.click.clientX
PASS 1.0 - 1.5: MouseEvent.click.clientY
PASS 1.0 - 1.5: MouseEvent.click.pageX
PASS 1.0 - 1.5: MouseEvent.click.pageY
PASS 1.0 - 1.5: MouseEvent.click.offsetX
PASS 1.0 - 1.5: MouseEvent.click.offsetY
FAIL -0.5 - 0.0: MouseEvent.click.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.click.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.click.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.click.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.click.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.auxclick.screenX
PASS 0.0 - 0.5: MouseEvent.auxclick.screenY
PASS 0.0 - 0.5: MouseEvent.auxclick.clientX
PASS 0.0 - 0.5: MouseEvent.auxclick.clientY
PASS 0.0 - 0.5: MouseEvent.auxclick.pageX
PASS 0.0 - 0.5: MouseEvent.auxclick.pageY
PASS 0.0 - 0.5: MouseEvent.auxclick.offsetX
PASS 0.0 - 0.5: MouseEvent.auxclick.offsetY
PASS 0.5 - 1.0: MouseEvent.auxclick.screenX
PASS 0.5 - 1.0: MouseEvent.auxclick.screenY
PASS 0.5 - 1.0: MouseEvent.auxclick.clientX
PASS 0.5 - 1.0: MouseEvent.auxclick.clientY
PASS 0.5 - 1.0: MouseEvent.auxclick.pageX
PASS 0.5 - 1.0: MouseEvent.auxclick.pageY
PASS 0.5 - 1.0: MouseEvent.auxclick.offsetX
PASS 0.5 - 1.0: MouseEvent.auxclick.offsetY
PASS 1.0 - 1.5: MouseEvent.auxclick.screenX
PASS 1.0 - 1.5: MouseEvent.auxclick.screenY
PASS 1.0 - 1.5: MouseEvent.auxclick.clientX
PASS 1.0 - 1.5: MouseEvent.auxclick.clientY
PASS 1.0 - 1.5: MouseEvent.auxclick.pageX
PASS 1.0 - 1.5: MouseEvent.auxclick.pageY
PASS 1.0 - 1.5: MouseEvent.auxclick.offsetX
PASS 1.0 - 1.5: MouseEvent.auxclick.offsetY
FAIL -0.5 - 0.0: MouseEvent.auxclick.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.auxclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.auxclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.auxclick.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.auxclick.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.contextmenu.screenX
PASS 0.0 - 0.5: MouseEvent.contextmenu.screenY
PASS 0.0 - 0.5: MouseEvent.contextmenu.clientX
PASS 0.0 - 0.5: MouseEvent.contextmenu.clientY
PASS 0.0 - 0.5: MouseEvent.contextmenu.pageX
PASS 0.0 - 0.5: MouseEvent.contextmenu.pageY
PASS 0.0 - 0.5: MouseEvent.contextmenu.offsetX
PASS 0.0 - 0.5: MouseEvent.contextmenu.offsetY
PASS 0.5 - 1.0: MouseEvent.contextmenu.screenX
PASS 0.5 - 1.0: MouseEvent.contextmenu.screenY
PASS 0.5 - 1.0: MouseEvent.contextmenu.clientX
PASS 0.5 - 1.0: MouseEvent.contextmenu.clientY
PASS 0.5 - 1.0: MouseEvent.contextmenu.pageX
PASS 0.5 - 1.0: MouseEvent.contextmenu.pageY
PASS 0.5 - 1.0: MouseEvent.contextmenu.offsetX
PASS 0.5 - 1.0: MouseEvent.contextmenu.offsetY
PASS 1.0 - 1.5: MouseEvent.contextmenu.screenX
PASS 1.0 - 1.5: MouseEvent.contextmenu.screenY
PASS 1.0 - 1.5: MouseEvent.contextmenu.clientX
PASS 1.0 - 1.5: MouseEvent.contextmenu.clientY
PASS 1.0 - 1.5: MouseEvent.contextmenu.pageX
PASS 1.0 - 1.5: MouseEvent.contextmenu.pageY
PASS 1.0 - 1.5: MouseEvent.contextmenu.offsetX
PASS 1.0 - 1.5: MouseEvent.contextmenu.offsetY
FAIL -0.5 - 0.0: MouseEvent.contextmenu.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.contextmenu.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.contextmenu.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.contextmenu.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.contextmenu.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mousedown.screenX
PASS 0.0 - 0.5: MouseEvent.mousedown.screenY
PASS 0.0 - 0.5: MouseEvent.mousedown.clientX
PASS 0.0 - 0.5: MouseEvent.mousedown.clientY
PASS 0.0 - 0.5: MouseEvent.mousedown.pageX
PASS 0.0 - 0.5: MouseEvent.mousedown.pageY
PASS 0.0 - 0.5: MouseEvent.mousedown.offsetX
PASS 0.0 - 0.5: MouseEvent.mousedown.offsetY
PASS 0.5 - 1.0: MouseEvent.mousedown.screenX
PASS 0.5 - 1.0: MouseEvent.mousedown.screenY
PASS 0.5 - 1.0: MouseEvent.mousedown.clientX
PASS 0.5 - 1.0: MouseEvent.mousedown.clientY
PASS 0.5 - 1.0: MouseEvent.mousedown.pageX
PASS 0.5 - 1.0: MouseEvent.mousedown.pageY
PASS 0.5 - 1.0: MouseEvent.mousedown.offsetX
PASS 0.5 - 1.0: MouseEvent.mousedown.offsetY
PASS 1.0 - 1.5: MouseEvent.mousedown.screenX
PASS 1.0 - 1.5: MouseEvent.mousedown.screenY
PASS 1.0 - 1.5: MouseEvent.mousedown.clientX
PASS 1.0 - 1.5: MouseEvent.mousedown.clientY
PASS 1.0 - 1.5: MouseEvent.mousedown.pageX
PASS 1.0 - 1.5: MouseEvent.mousedown.pageY
PASS 1.0 - 1.5: MouseEvent.mousedown.offsetX
PASS 1.0 - 1.5: MouseEvent.mousedown.offsetY
FAIL -0.5 - 0.0: MouseEvent.mousedown.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousedown.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousedown.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mousedown.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousedown.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mousemove.screenX
PASS 0.0 - 0.5: MouseEvent.mousemove.screenY
PASS 0.0 - 0.5: MouseEvent.mousemove.clientX
PASS 0.0 - 0.5: MouseEvent.mousemove.clientY
PASS 0.0 - 0.5: MouseEvent.mousemove.pageX
PASS 0.0 - 0.5: MouseEvent.mousemove.pageY
PASS 0.0 - 0.5: MouseEvent.mousemove.offsetX
PASS 0.0 - 0.5: MouseEvent.mousemove.offsetY
PASS 0.5 - 1.0: MouseEvent.mousemove.screenX
PASS 0.5 - 1.0: MouseEvent.mousemove.screenY
PASS 0.5 - 1.0: MouseEvent.mousemove.clientX
PASS 0.5 - 1.0: MouseEvent.mousemove.clientY
PASS 0.5 - 1.0: MouseEvent.mousemove.pageX
PASS 0.5 - 1.0: MouseEvent.mousemove.pageY
PASS 0.5 - 1.0: MouseEvent.mousemove.offsetX
PASS 0.5 - 1.0: MouseEvent.mousemove.offsetY
PASS 1.0 - 1.5: MouseEvent.mousemove.screenX
PASS 1.0 - 1.5: MouseEvent.mousemove.screenY
PASS 1.0 - 1.5: MouseEvent.mousemove.clientX
PASS 1.0 - 1.5: MouseEvent.mousemove.clientY
PASS 1.0 - 1.5: MouseEvent.mousemove.pageX
PASS 1.0 - 1.5: MouseEvent.mousemove.pageY
PASS 1.0 - 1.5: MouseEvent.mousemove.offsetX
PASS 1.0 - 1.5: MouseEvent.mousemove.offsetY
FAIL -0.5 - 0.0: MouseEvent.mousemove.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mousemove.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mousemove.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mousemove.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mousemove.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mouseup.screenX
PASS 0.0 - 0.5: MouseEvent.mouseup.screenY
PASS 0.0 - 0.5: MouseEvent.mouseup.clientX
PASS 0.0 - 0.5: MouseEvent.mouseup.clientY
PASS 0.0 - 0.5: MouseEvent.mouseup.pageX
PASS 0.0 - 0.5: MouseEvent.mouseup.pageY
PASS 0.0 - 0.5: MouseEvent.mouseup.offsetX
PASS 0.0 - 0.5: MouseEvent.mouseup.offsetY
PASS 0.5 - 1.0: MouseEvent.mouseup.screenX
PASS 0.5 - 1.0: MouseEvent.mouseup.screenY
PASS 0.5 - 1.0: MouseEvent.mouseup.clientX
PASS 0.5 - 1.0: MouseEvent.mouseup.clientY
PASS 0.5 - 1.0: MouseEvent.mouseup.pageX
PASS 0.5 - 1.0: MouseEvent.mouseup.pageY
PASS 0.5 - 1.0: MouseEvent.mouseup.offsetX
PASS 0.5 - 1.0: MouseEvent.mouseup.offsetY
PASS 1.0 - 1.5: MouseEvent.mouseup.screenX
PASS 1.0 - 1.5: MouseEvent.mouseup.screenY
PASS 1.0 - 1.5: MouseEvent.mouseup.clientX
PASS 1.0 - 1.5: MouseEvent.mouseup.clientY
PASS 1.0 - 1.5: MouseEvent.mouseup.pageX
PASS 1.0 - 1.5: MouseEvent.mouseup.pageY
PASS 1.0 - 1.5: MouseEvent.mouseup.offsetX
PASS 1.0 - 1.5: MouseEvent.mouseup.offsetY
FAIL -0.5 - 0.0: MouseEvent.mouseup.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseup.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseup.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mouseup.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseup.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mouseover.screenX
PASS 0.0 - 0.5: MouseEvent.mouseover.screenY
PASS 0.0 - 0.5: MouseEvent.mouseover.clientX
PASS 0.0 - 0.5: MouseEvent.mouseover.clientY
PASS 0.0 - 0.5: MouseEvent.mouseover.pageX
PASS 0.0 - 0.5: MouseEvent.mouseover.pageY
PASS 0.0 - 0.5: MouseEvent.mouseover.offsetX
PASS 0.0 - 0.5: MouseEvent.mouseover.offsetY
PASS 0.5 - 1.0: MouseEvent.mouseover.screenX
PASS 0.5 - 1.0: MouseEvent.mouseover.screenY
PASS 0.5 - 1.0: MouseEvent.mouseover.clientX
PASS 0.5 - 1.0: MouseEvent.mouseover.clientY
PASS 0.5 - 1.0: MouseEvent.mouseover.pageX
PASS 0.5 - 1.0: MouseEvent.mouseover.pageY
PASS 0.5 - 1.0: MouseEvent.mouseover.offsetX
PASS 0.5 - 1.0: MouseEvent.mouseover.offsetY
PASS 1.0 - 1.5: MouseEvent.mouseover.screenX
PASS 1.0 - 1.5: MouseEvent.mouseover.screenY
PASS 1.0 - 1.5: MouseEvent.mouseover.clientX
PASS 1.0 - 1.5: MouseEvent.mouseover.clientY
PASS 1.0 - 1.5: MouseEvent.mouseover.pageX
PASS 1.0 - 1.5: MouseEvent.mouseover.pageY
PASS 1.0 - 1.5: MouseEvent.mouseover.offsetX
PASS 1.0 - 1.5: MouseEvent.mouseover.offsetY
FAIL -0.5 - 0.0: MouseEvent.mouseover.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseover.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseover.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mouseover.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseover.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mouseout.screenX
PASS 0.0 - 0.5: MouseEvent.mouseout.screenY
PASS 0.0 - 0.5: MouseEvent.mouseout.clientX
PASS 0.0 - 0.5: MouseEvent.mouseout.clientY
PASS 0.0 - 0.5: MouseEvent.mouseout.pageX
PASS 0.0 - 0.5: MouseEvent.mouseout.pageY
PASS 0.0 - 0.5: MouseEvent.mouseout.offsetX
PASS 0.0 - 0.5: MouseEvent.mouseout.offsetY
PASS 0.5 - 1.0: MouseEvent.mouseout.screenX
PASS 0.5 - 1.0: MouseEvent.mouseout.screenY
PASS 0.5 - 1.0: MouseEvent.mouseout.clientX
PASS 0.5 - 1.0: MouseEvent.mouseout.clientY
PASS 0.5 - 1.0: MouseEvent.mouseout.pageX
PASS 0.5 - 1.0: MouseEvent.mouseout.pageY
PASS 0.5 - 1.0: MouseEvent.mouseout.offsetX
PASS 0.5 - 1.0: MouseEvent.mouseout.offsetY
PASS 1.0 - 1.5: MouseEvent.mouseout.screenX
PASS 1.0 - 1.5: MouseEvent.mouseout.screenY
PASS 1.0 - 1.5: MouseEvent.mouseout.clientX
PASS 1.0 - 1.5: MouseEvent.mouseout.clientY
PASS 1.0 - 1.5: MouseEvent.mouseout.pageX
PASS 1.0 - 1.5: MouseEvent.mouseout.pageY
PASS 1.0 - 1.5: MouseEvent.mouseout.offsetX
PASS 1.0 - 1.5: MouseEvent.mouseout.offsetY
FAIL -0.5 - 0.0: MouseEvent.mouseout.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseout.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseout.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mouseout.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseout.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mouseenter.screenX
PASS 0.0 - 0.5: MouseEvent.mouseenter.screenY
PASS 0.0 - 0.5: MouseEvent.mouseenter.clientX
PASS 0.0 - 0.5: MouseEvent.mouseenter.clientY
PASS 0.0 - 0.5: MouseEvent.mouseenter.pageX
PASS 0.0 - 0.5: MouseEvent.mouseenter.pageY
PASS 0.0 - 0.5: MouseEvent.mouseenter.offsetX
PASS 0.0 - 0.5: MouseEvent.mouseenter.offsetY
PASS 0.5 - 1.0: MouseEvent.mouseenter.screenX
PASS 0.5 - 1.0: MouseEvent.mouseenter.screenY
PASS 0.5 - 1.0: MouseEvent.mouseenter.clientX
PASS 0.5 - 1.0: MouseEvent.mouseenter.clientY
PASS 0.5 - 1.0: MouseEvent.mouseenter.pageX
PASS 0.5 - 1.0: MouseEvent.mouseenter.pageY
PASS 0.5 - 1.0: MouseEvent.mouseenter.offsetX
PASS 0.5 - 1.0: MouseEvent.mouseenter.offsetY
PASS 1.0 - 1.5: MouseEvent.mouseenter.screenX
PASS 1.0 - 1.5: MouseEvent.mouseenter.screenY
PASS 1.0 - 1.5: MouseEvent.mouseenter.clientX
PASS 1.0 - 1.5: MouseEvent.mouseenter.clientY
PASS 1.0 - 1.5: MouseEvent.mouseenter.pageX
PASS 1.0 - 1.5: MouseEvent.mouseenter.pageY
PASS 1.0 - 1.5: MouseEvent.mouseenter.offsetX
PASS 1.0 - 1.5: MouseEvent.mouseenter.offsetY
FAIL -0.5 - 0.0: MouseEvent.mouseenter.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseenter.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseenter.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mouseenter.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseenter.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.mouseleave.screenX
PASS 0.0 - 0.5: MouseEvent.mouseleave.screenY
PASS 0.0 - 0.5: MouseEvent.mouseleave.clientX
PASS 0.0 - 0.5: MouseEvent.mouseleave.clientY
PASS 0.0 - 0.5: MouseEvent.mouseleave.pageX
PASS 0.0 - 0.5: MouseEvent.mouseleave.pageY
PASS 0.0 - 0.5: MouseEvent.mouseleave.offsetX
PASS 0.0 - 0.5: MouseEvent.mouseleave.offsetY
PASS 0.5 - 1.0: MouseEvent.mouseleave.screenX
PASS 0.5 - 1.0: MouseEvent.mouseleave.screenY
PASS 0.5 - 1.0: MouseEvent.mouseleave.clientX
PASS 0.5 - 1.0: MouseEvent.mouseleave.clientY
PASS 0.5 - 1.0: MouseEvent.mouseleave.pageX
PASS 0.5 - 1.0: MouseEvent.mouseleave.pageY
PASS 0.5 - 1.0: MouseEvent.mouseleave.offsetX
PASS 0.5 - 1.0: MouseEvent.mouseleave.offsetY
PASS 1.0 - 1.5: MouseEvent.mouseleave.screenX
PASS 1.0 - 1.5: MouseEvent.mouseleave.screenY
PASS 1.0 - 1.5: MouseEvent.mouseleave.clientX
PASS 1.0 - 1.5: MouseEvent.mouseleave.clientY
PASS 1.0 - 1.5: MouseEvent.mouseleave.pageX
PASS 1.0 - 1.5: MouseEvent.mouseleave.pageY
PASS 1.0 - 1.5: MouseEvent.mouseleave.offsetX
PASS 1.0 - 1.5: MouseEvent.mouseleave.offsetY
FAIL -0.5 - 0.0: MouseEvent.mouseleave.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.mouseleave.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.mouseleave.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.mouseleave.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.mouseleave.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: MouseEvent.dblclick.screenX
PASS 0.0 - 0.5: MouseEvent.dblclick.screenY
PASS 0.0 - 0.5: MouseEvent.dblclick.clientX
PASS 0.0 - 0.5: MouseEvent.dblclick.clientY
PASS 0.0 - 0.5: MouseEvent.dblclick.pageX
PASS 0.0 - 0.5: MouseEvent.dblclick.pageY
PASS 0.0 - 0.5: MouseEvent.dblclick.offsetX
PASS 0.0 - 0.5: MouseEvent.dblclick.offsetY
PASS 0.5 - 1.0: MouseEvent.dblclick.screenX
PASS 0.5 - 1.0: MouseEvent.dblclick.screenY
PASS 0.5 - 1.0: MouseEvent.dblclick.clientX
PASS 0.5 - 1.0: MouseEvent.dblclick.clientY
PASS 0.5 - 1.0: MouseEvent.dblclick.pageX
PASS 0.5 - 1.0: MouseEvent.dblclick.pageY
PASS 0.5 - 1.0: MouseEvent.dblclick.offsetX
PASS 0.5 - 1.0: MouseEvent.dblclick.offsetY
PASS 1.0 - 1.5: MouseEvent.dblclick.screenX
PASS 1.0 - 1.5: MouseEvent.dblclick.screenY
PASS 1.0 - 1.5: MouseEvent.dblclick.clientX
PASS 1.0 - 1.5: MouseEvent.dblclick.clientY
PASS 1.0 - 1.5: MouseEvent.dblclick.pageX
PASS 1.0 - 1.5: MouseEvent.dblclick.pageY
PASS 1.0 - 1.5: MouseEvent.dblclick.offsetX
PASS 1.0 - 1.5: MouseEvent.dblclick.offsetY
FAIL -0.5 - 0.0: MouseEvent.dblclick.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: MouseEvent.dblclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: MouseEvent.dblclick.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: MouseEvent.dblclick.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: MouseEvent.dblclick.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: DragEvent.dragstart.screenX
PASS 0.0 - 0.5: DragEvent.dragstart.screenY
PASS 0.0 - 0.5: DragEvent.dragstart.clientX
PASS 0.0 - 0.5: DragEvent.dragstart.clientY
PASS 0.0 - 0.5: DragEvent.dragstart.pageX
PASS 0.0 - 0.5: DragEvent.dragstart.pageY
PASS 0.0 - 0.5: DragEvent.dragstart.offsetX
PASS 0.0 - 0.5: DragEvent.dragstart.offsetY
PASS 0.5 - 1.0: DragEvent.dragstart.screenX
PASS 0.5 - 1.0: DragEvent.dragstart.screenY
PASS 0.5 - 1.0: DragEvent.dragstart.clientX
PASS 0.5 - 1.0: DragEvent.dragstart.clientY
PASS 0.5 - 1.0: DragEvent.dragstart.pageX
PASS 0.5 - 1.0: DragEvent.dragstart.pageY
PASS 0.5 - 1.0: DragEvent.dragstart.offsetX
PASS 0.5 - 1.0: DragEvent.dragstart.offsetY
PASS 1.0 - 1.5: DragEvent.dragstart.screenX
PASS 1.0 - 1.5: DragEvent.dragstart.screenY
PASS 1.0 - 1.5: DragEvent.dragstart.clientX
PASS 1.0 - 1.5: DragEvent.dragstart.clientY
PASS 1.0 - 1.5: DragEvent.dragstart.pageX
PASS 1.0 - 1.5: DragEvent.dragstart.pageY
PASS 1.0 - 1.5: DragEvent.dragstart.offsetX
PASS 1.0 - 1.5: DragEvent.dragstart.offsetY
FAIL -0.5 - 0.0: DragEvent.dragstart.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragstart.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragstart.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: DragEvent.dragstart.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragstart.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: DragEvent.dragover.screenX
PASS 0.0 - 0.5: DragEvent.dragover.screenY
PASS 0.0 - 0.5: DragEvent.dragover.clientX
PASS 0.0 - 0.5: DragEvent.dragover.clientY
PASS 0.0 - 0.5: DragEvent.dragover.pageX
PASS 0.0 - 0.5: DragEvent.dragover.pageY
PASS 0.0 - 0.5: DragEvent.dragover.offsetX
PASS 0.0 - 0.5: DragEvent.dragover.offsetY
PASS 0.5 - 1.0: DragEvent.dragover.screenX
PASS 0.5 - 1.0: DragEvent.dragover.screenY
PASS 0.5 - 1.0: DragEvent.dragover.clientX
PASS 0.5 - 1.0: DragEvent.dragover.clientY
PASS 0.5 - 1.0: DragEvent.dragover.pageX
PASS 0.5 - 1.0: DragEvent.dragover.pageY
PASS 0.5 - 1.0: DragEvent.dragover.offsetX
PASS 0.5 - 1.0: DragEvent.dragover.offsetY
PASS 1.0 - 1.5: DragEvent.dragover.screenX
PASS 1.0 - 1.5: DragEvent.dragover.screenY
PASS 1.0 - 1.5: DragEvent.dragover.clientX
PASS 1.0 - 1.5: DragEvent.dragover.clientY
PASS 1.0 - 1.5: DragEvent.dragover.pageX
PASS 1.0 - 1.5: DragEvent.dragover.pageY
PASS 1.0 - 1.5: DragEvent.dragover.offsetX
PASS 1.0 - 1.5: DragEvent.dragover.offsetY
FAIL -0.5 - 0.0: DragEvent.dragover.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragover.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragover.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: DragEvent.dragover.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragover.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: DragEvent.dragend.screenX
PASS 0.0 - 0.5: DragEvent.dragend.screenY
PASS 0.0 - 0.5: DragEvent.dragend.clientX
PASS 0.0 - 0.5: DragEvent.dragend.clientY
PASS 0.0 - 0.5: DragEvent.dragend.pageX
PASS 0.0 - 0.5: DragEvent.dragend.pageY
PASS 0.0 - 0.5: DragEvent.dragend.offsetX
PASS 0.0 - 0.5: DragEvent.dragend.offsetY
PASS 0.5 - 1.0: DragEvent.dragend.screenX
PASS 0.5 - 1.0: DragEvent.dragend.screenY
PASS 0.5 - 1.0: DragEvent.dragend.clientX
PASS 0.5 - 1.0: DragEvent.dragend.clientY
PASS 0.5 - 1.0: DragEvent.dragend.pageX
PASS 0.5 - 1.0: DragEvent.dragend.pageY
PASS 0.5 - 1.0: DragEvent.dragend.offsetX
PASS 0.5 - 1.0: DragEvent.dragend.offsetY
PASS 1.0 - 1.5: DragEvent.dragend.screenX
PASS 1.0 - 1.5: DragEvent.dragend.screenY
PASS 1.0 - 1.5: DragEvent.dragend.clientX
PASS 1.0 - 1.5: DragEvent.dragend.clientY
PASS 1.0 - 1.5: DragEvent.dragend.pageX
PASS 1.0 - 1.5: DragEvent.dragend.pageY
PASS 1.0 - 1.5: DragEvent.dragend.offsetX
PASS 1.0 - 1.5: DragEvent.dragend.offsetY
FAIL -0.5 - 0.0: DragEvent.dragend.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.dragend.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.dragend.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: DragEvent.dragend.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.dragend.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: DragEvent.drop.screenX
PASS 0.0 - 0.5: DragEvent.drop.screenY
PASS 0.0 - 0.5: DragEvent.drop.clientX
PASS 0.0 - 0.5: DragEvent.drop.clientY
PASS 0.0 - 0.5: DragEvent.drop.pageX
PASS 0.0 - 0.5: DragEvent.drop.pageY
PASS 0.0 - 0.5: DragEvent.drop.offsetX
PASS 0.0 - 0.5: DragEvent.drop.offsetY
PASS 0.5 - 1.0: DragEvent.drop.screenX
PASS 0.5 - 1.0: DragEvent.drop.screenY
PASS 0.5 - 1.0: DragEvent.drop.clientX
PASS 0.5 - 1.0: DragEvent.drop.clientY
PASS 0.5 - 1.0: DragEvent.drop.pageX
PASS 0.5 - 1.0: DragEvent.drop.pageY
PASS 0.5 - 1.0: DragEvent.drop.offsetX
PASS 0.5 - 1.0: DragEvent.drop.offsetY
PASS 1.0 - 1.5: DragEvent.drop.screenX
PASS 1.0 - 1.5: DragEvent.drop.screenY
PASS 1.0 - 1.5: DragEvent.drop.clientX
PASS 1.0 - 1.5: DragEvent.drop.clientY
PASS 1.0 - 1.5: DragEvent.drop.pageX
PASS 1.0 - 1.5: DragEvent.drop.pageY
PASS 1.0 - 1.5: DragEvent.drop.offsetX
PASS 1.0 - 1.5: DragEvent.drop.offsetY
FAIL -0.5 - 0.0: DragEvent.drop.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: DragEvent.drop.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: DragEvent.drop.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: DragEvent.drop.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: DragEvent.drop.offsetY assert_equals: expected -2 but got -1
PASS 0.0 - 0.5: WheelEvent.wheel.screenX
PASS 0.0 - 0.5: WheelEvent.wheel.screenY
PASS 0.0 - 0.5: WheelEvent.wheel.clientX
PASS 0.0 - 0.5: WheelEvent.wheel.clientY
PASS 0.0 - 0.5: WheelEvent.wheel.pageX
PASS 0.0 - 0.5: WheelEvent.wheel.pageY
PASS 0.0 - 0.5: WheelEvent.wheel.offsetX
PASS 0.0 - 0.5: WheelEvent.wheel.offsetY
PASS 0.5 - 1.0: WheelEvent.wheel.screenX
PASS 0.5 - 1.0: WheelEvent.wheel.screenY
PASS 0.5 - 1.0: WheelEvent.wheel.clientX
PASS 0.5 - 1.0: WheelEvent.wheel.clientY
PASS 0.5 - 1.0: WheelEvent.wheel.pageX
PASS 0.5 - 1.0: WheelEvent.wheel.pageY
PASS 0.5 - 1.0: WheelEvent.wheel.offsetX
PASS 0.5 - 1.0: WheelEvent.wheel.offsetY
PASS 1.0 - 1.5: WheelEvent.wheel.screenX
PASS 1.0 - 1.5: WheelEvent.wheel.screenY
PASS 1.0 - 1.5: WheelEvent.wheel.clientX
PASS 1.0 - 1.5: WheelEvent.wheel.clientY
PASS 1.0 - 1.5: WheelEvent.wheel.pageX
PASS 1.0 - 1.5: WheelEvent.wheel.pageY
PASS 1.0 - 1.5: WheelEvent.wheel.offsetX
PASS 1.0 - 1.5: WheelEvent.wheel.offsetY
FAIL -0.5 - 0.0: WheelEvent.wheel.screenX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.screenY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.clientX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.clientY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.pageX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.pageY assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.offsetX assert_equals: expected -1 but got 0
FAIL -0.5 - 0.0: WheelEvent.wheel.offsetY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.screenX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.screenY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.clientX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.clientY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.pageX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.pageY assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.offsetX assert_equals: expected -1 but got 0
FAIL -1.0 - -0.5: WheelEvent.wheel.offsetY assert_equals: expected -1 but got 0
FAIL -1.5 - -1.0: WheelEvent.wheel.screenX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.screenY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.clientX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.clientY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.pageX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.pageY assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.offsetX assert_equals: expected -2 but got -1
FAIL -1.5 - -1.0: WheelEvent.wheel.offsetY assert_equals: expected -2 but got -1
PASS PointerEvent.pointerdown.width
PASS PointerEvent.pointerdown.height
PASS PointerEvent.pointerup.width
PASS PointerEvent.pointerup.height
PASS PointerEvent.pointermove.width
PASS PointerEvent.pointermove.height
PASS PointerEvent.pointercancel.width
PASS PointerEvent.pointercancel.height
PASS PointerEvent.gotpointercapture.width
PASS PointerEvent.gotpointercapture.height
PASS PointerEvent.lostpointercapture.width
PASS PointerEvent.lostpointercapture.height
PASS PointerEvent.pointerrawupdate.width
PASS PointerEvent.pointerrawupdate.height
PASS PointerEvent.pointerover.width
PASS PointerEvent.pointerover.height
PASS PointerEvent.pointerout.width
PASS PointerEvent.pointerout.height
PASS PointerEvent.pointerenter.width
PASS PointerEvent.pointerenter.height
PASS PointerEvent.pointerleave.width
PASS PointerEvent.pointerleave.height
PASS PointerEvent.foo.width
PASS PointerEvent.foo.height
